add Auth, Login, Search, 404

+++ app/views/posts/show.blade.php
@@ -8,16 +8,21 @@ Reader
     <div class="jumbotron">
         <h1>{{{ $post->title }}}</h1>
         <h3>{{{ $post->tldr }}}</h3>
+        <h5>{{{ 'Posted By: ' . $post->user->first_name . ' ' . $post->user->last_name }}}</h5>
         <h6>{{{ $post->created_at->setTimezone('America/Chicago')->format('l, F jS Y @ h:i: A') }}}</h6>
         <p>{{{ $post->body }}}</p>
     </div>
     {{-- @if(Auth::check() $$ Auth::id() == $post->user_id) --}}
-    <a class="btn btn-primary" href="{{{ action('PostsController@index') }}}"><span class="glyphicon glyphicon-fast-backward"> Back to All</a>
-    <a class="btn btn-warning" href="{{{ action('PostsController@edit', $post->id) }}}"><span class="glyphicon glyphicon-wrench"></span> Edit This Post</a>
-    <button id="delete" class="btn btn-danger" ><span class="glyphicon glyphicon-remove"> Delete</button>
-    {{ Form::open(array('action' => array('PostsController@destroy', $post->id), 'method' => 'DELETE', 'id' => 'formDelete')) }}
-    {{ Form::close() }}
-    {{-- @endif --}}
+    <a href="{{{ action('PostsController@index') }}}"><button class="btn btn-primary" ><span class="glyphicon glyphicon-fast-backward"></span> Back to All</button></a>
+
+    @if(Auth::check() && Auth::id() == $post->user_id)
+    
+        <a href="{{{ action('PostsController@edit', $post->id) }}}"><button class="btn btn-warning" ><span class="glyphicon glyphicon-wrench"></span> Edit This Post</button></a>
+
+        <button id="delete" class="btn btn-danger" ><span class="glyphicon glyphicon-remove"></span> Delete</button>
+        {{ Form::open(array('action' => array('PostsController@destroy', $post->id), 'method' => 'DELETE', 'id' => 'formDelete')) }}
+        {{ Form::close() }}
+    @endif 
 @stop
 
 @section('script')
